+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Enhancements
 
+* Add `A.Vector.flat_map/2`
 * `A.OrdMap.drop/2` will only try to period rebuild once after dropping all keys
 
 ### Bug fixes